# graphql-go-crud/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Graphql golang API",
        "title": "graphql-go-crud",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/goengine/mylist": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get list of mylist data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mylist"
                ],
                "summary": "Get list of mylist data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Mytype"
                        }
                    }
                }
            }
        },
        "/goengine/skills": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get list of skill",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skills"
                ],
                "summary": "Get list of skill",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.SkilltypeList"
                            }
                        }
                    }
                }
            }
        },
        "/token": {
            "post": {
                "description": "Get token to access application",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "Get token to access application",
                "parameters": [
                    {
                        "description": "Credentials info",
                        "name": "Credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Credentials"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Credentials": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "main.Mytype": {
            "type": "object",
            "properties": {
                "float32": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "info": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "main.SkilltypeList": {
            "type": "object",
            "properties": {
                "skills": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/skilldetailsmodel.SkillType"
                    }
                }
            }
        },
        "main.Token": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "skilldetailsmodel.SkillType": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                },
                "skill_id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}